+++ libft/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: kturken <kturken@student.42kocaeli.com.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/12 14:19:06 by kturken           #+#    #+#             */
-/*   Updated: 2022/12/19 13:47:52 by kturken          ###   ########.fr       */
+/*   Updated: 2022/12/19 15:22:11 by kturken          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	char	*str;
-	int		i;
+	char		*str;
+	size_t		i;
 
 	i = 0;
 	str = (char *) malloc(len + 1);
